+++ src/Application/Auth/Commands/SignUp/SignUpHandler.cs
@@ -2,26 +2,26 @@
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Application.Auth.Commands.Login;
 using Application.Common.Exceptions;
 using Application.Common.Interfaces;
 using Application.Common.Models;
 using MediatR;
 
-namespace Application.Auth.RequestHandlers
+namespace Application.Auth.Commands.SignUp
 {
-    public class LoginHandler : IRequestHandler<LoginCommand, AuthenticationResponse>
+    public class SignUpHandler : IRequestHandler<SignUpCommand, AuthenticationResponse>
     {
         private readonly IIdentityService _identityService;
 
-        public LoginHandler(IIdentityService identityService)
+        public SignUpHandler(IIdentityService identityService)
         {
             _identityService = identityService;
         }
 
-        public async Task<AuthenticationResponse> Handle(LoginCommand request, CancellationToken cancellationToken)
+
+        public async Task<AuthenticationResponse> Handle(SignUpCommand request, CancellationToken cancellationToken)
         {
-            var result = await _identityService.Login(request.UserLoginRequest);
+            var result = await _identityService.Register(request.UserSignUpRequest);
 
             if (result.ErrorMessages.Any())
             {
