+++ src/Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantHandler.cs
@@ -2,10 +2,9 @@
 using System.Threading.Tasks;
 using Application.Common.Exceptions;
 using Application.Common.Interfaces;
-using Application.Restaurants.Commands.DeleteRestaurant;
 using MediatR;
 
-namespace Application.Restaurants.RequestHandlers
+namespace Application.Restaurants.Commands.DeleteRestaurant
 {
     public class DeleteRestaurantHandler : IRequestHandler<DeleteRestaurantCommand>
     {
